backend/requirements.txt
geocoder
fastapi
pydantic
pydantic_ai
groq
python-dotenv
uvicorn